k8s/apps/terraria/files/discord-bot/pyproject.toml
[project]
name = "terraria-discord-bot"
version = "0.1.0"
description = "Discord bot for managing a Terraria server on Kubernetes"
requires-python = ">=3.12"
dependencies = [
    "pynacl>=1.5.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
    "integration: marks tests as integration tests (deselected by default)",
]
addopts = "-m 'not integration'"
